requirements.txt
openai==1.30.3
requests==2.32.2
python-dotenv==1.0.1
pydub==0.25.1
